test: move test to js-sourcemap
