+++ playground/js-sourcemap/index.html
@@ -4,6 +4,6 @@
 </div>
 
 <script type="module" src="./foo.js"></script>
-<script type="module" src="./bar.ts"></script>
+<script type="module" src="./bar.js"></script>
 <script type="module" src="./after-preload-dynamic.js"></script>
 <script type="module" src="./with-multiline-import.ts"></script>
