+++ playground/js-sourcemap/plugin-foo.js
