+++ playground/js-sourcemap/vite.config.js
@@ -1,8 +1,12 @@
 import { defineConfig } from 'vite'
 import transformFooWithInlineSourceMap from './foo-with-sourcemap-plugin'
+import { transformZooWithSourcemapPlugin } from './zoo-with-sourcemap-plugin'
 
 export default defineConfig({
-  plugins: [transformFooWithInlineSourceMap()],
+  plugins: [
+    transformFooWithInlineSourceMap(),
+    transformZooWithSourcemapPlugin(),
+  ],
   build: {
     sourcemap: true,
     rollupOptions: {
