+++ playground/js-sourcemap/zoo.js
@@ -1,0 +1,1 @@
+export const zoo = 'zoo'
