--- playground/plugin-sourcemap/index.html
@@ -1,5 +1,0 @@
-<div class="wrapper">
-  <h1>Plugin Sourcemap</h1>
-</div>
-
-<script type="module" src="./foo.js"></script>
