--- playground/plugin-sourcemap/package.json
@@ -1,15 +1,0 @@
-{
-  "name": "@vitejs/test-plugin-sourcemap",
-  "private": true,
-  "version": "0.0.0",
-  "type": "module",
-  "scripts": {
-    "dev": "vite",
-    "build": "vite build",
-    "debug": "node --inspect-brk ../../packages/vite/bin/vite",
-    "preview": "vite preview"
-  },
-  "devDependencies": {
-    "magic-string": "^0.30.3"
-  }
-}
